fix(hosting): correct internal route path, gate card on secret, split stats by currency
Address fresh review of the reworked commit:
- Move the activation handler to its own /v1/internal group (routes/internal.ts) so the real path matches the documented /v1/internal/activate (it was /v1/payments/internal/activate under the payments mount -> ePoints would 404 and paid tenants stay uncredited).
- /methods only reports card.enabled when HOSTING_INTERNAL_SECRET is set, so the UI never offers card when the activation call can only 403.
- Split /stats + the payment_stats view amounts by currency (HBD vs USD) so card revenue can't corrupt HBD totals; migration does CREATE OR REPLACE VIEW.
