+++ apps/self-hosted/hosting/api/src/routes/payments.ts
@@ -8,21 +8,15 @@ import { authMiddleware, adminMiddleware } from '../middleware/auth';
 
 export const paymentRoutes = new Hono();
 
-// Constant-time check of the shared service-to-service secret (ePoints -> hosting).
-// Fails CLOSED when the secret is unset, so a misconfigured deploy cannot activate blogs.
-function internalSecretOk(provided: string | undefined): boolean {
-  const secret = process.env.HOSTING_INTERNAL_SECRET || '';
-  const given = provided || '';
-  if (secret.length === 0 || given.length !== secret.length) return false;
-  let diff = 0;
-  for (let i = 0; i < secret.length; i++) diff |= given.charCodeAt(i) ^ secret.charCodeAt(i);
-  return diff === 0;
-}
-
 // GET /v1/payments/methods - which rails are available + pricing (for the signup UI)
 paymentRoutes.get('/methods', async (c) => {
   const monthlyHbd = process.env.MONTHLY_PRICE_HBD || '2.000';
   const facilitator = process.env.X402_FACILITATOR_URL || '';
+  // Card requires the internal secret (ePoints -> hosting activation). Without it, a paid
+  // card order can only get a 403 from /v1/internal/activate, so never advertise the option.
+  const cardEnabled =
+    (process.env.HOSTING_CARD_ENABLED || 'true') === 'true' &&
+    (process.env.HOSTING_INTERNAL_SECRET || '').length > 0;
   return c.json({
     hbd: {
       enabled: true,
@@ -37,104 +31,10 @@ paymentRoutes.get('/methods', async (c) => {
     // the web via the shared Elements flow); this flag only tells the UI whether to
     // offer the option.
     card: {
-      enabled: (process.env.HOSTING_CARD_ENABLED || 'true') === 'true',
+      enabled: cardEnabled,
       monthlyUsdCents: parseInt(process.env.HOSTING_CARD_USD_CENTS || '200', 10),
     },
   });
-});
-
-// POST /v1/internal/activate - service-to-service activation (NOT public).
-// Called by ePoints' fulfillment after a paid card order settles (ePoints is the single
-// Stripe brain; hosting never sees card data or Stripe keys). Guarded by a shared secret.
-// Idempotent + atomic: dedup on the order id so a fulfillment retry can never double-extend,
-// and the payment record + activation commit together so a crash can't leave a paid tenant
-// uncredited. Returns 200 on success or benign replay; 404 (permanent) if the tenant is gone
-// so ePoints FAILs the order; 5xx (transient) so ePoints leaves it VERIFIED and retries.
-paymentRoutes.post('/internal/activate', async (c) => {
-  if (!internalSecretOk(c.req.header('x-internal-secret'))) {
-    return c.json({ error: 'forbidden' }, 403);
-  }
-
-  let body: any;
-  try {
-    body = await c.req.json();
-  } catch {
-    return c.json({ error: 'invalid_body' }, 400);
-  }
-
-  const username = String(body?.username || '').toLowerCase();
-  // Strict: reject a missing/garbage/out-of-range term rather than silently granting 1 month.
-  const monthsRaw = parseInt(body?.months, 10);
-  const months = Number.isInteger(monthsRaw) ? monthsRaw : 0;
-  const orderId = String(body?.order_id || '').trim();
-  const amountUsd = typeof body?.amount_usd === 'number' ? body.amount_usd : 0;
-
-  if (!/^[a-z][a-z0-9.-]{2,15}$/.test(username) || months < 1 || months > 24 || !orderId) {
-    return c.json({ error: 'invalid_request' }, 400);
-  }
-
-  try {
-    const result = await db.transaction<{ status: 200 | 404; duplicate?: boolean; expiresAt?: Date }>(
-      async (client) => {
-        // Lock the tenant row (serializes retries) and confirm it still exists.
-        const t = await client.query(
-          `SELECT id, subscription_started_at, subscription_expires_at
-             FROM tenants WHERE username = $1 FOR UPDATE`,
-          [username]
-        );
-        if (t.rowCount === 0) {
-          return { status: 404 };
-        }
-        const tenant = t.rows[0];
-
-        // Claim the order id. If already recorded, this is a replay -> no re-credit.
-        const claim = await client.query(
-          `INSERT INTO payments
-             (tenant_id, trx_id, block_num, from_account, amount, currency, memo, status, months_credited)
-           VALUES ($1, $2, 0, 'stripe', $3, 'USD', $4, 'processed', $5)
-           ON CONFLICT (trx_id) DO NOTHING
-           RETURNING id`,
-          [tenant.id, orderId, amountUsd, `blog:${username}:${months}`, months]
-        );
-        if (claim.rowCount === 0) {
-          return { status: 200, duplicate: true };
-        }
-
-        // Extend from the later of now / current expiry (same rule as activateSubscription).
-        const now = new Date();
-        const currentExpiry = tenant.subscription_expires_at
-          ? new Date(tenant.subscription_expires_at)
-          : now;
-        const base = currentExpiry > now ? currentExpiry : now;
-        const newExpiry = new Date(base);
-        newExpiry.setMonth(newExpiry.getMonth() + months);
-        const startedAt = tenant.subscription_started_at || now;
-
-        await client.query(
-          `UPDATE tenants
-             SET subscription_status = 'active',
-                 subscription_started_at = $2,
-                 subscription_expires_at = $3,
-                 updated_at = NOW()
-           WHERE id = $1`,
-          [tenant.id, startedAt, newExpiry]
-        );
-        await client.query(
-          `UPDATE payments SET processed_at = NOW(), subscription_extended_to = $2 WHERE trx_id = $1`,
-          [orderId, newExpiry]
-        );
-        return { status: 200, expiresAt: newExpiry };
-      }
-    );
-
-    if (result.status === 404) {
-      return c.json({ error: 'tenant_not_found' }, 404);
-    }
-    return c.json({ activated: true, duplicate: !!result.duplicate, expiresAt: result.expiresAt }, 200);
-  } catch (e) {
-    console.error('[internal/activate] error:', (e as Error).message);
-    return c.json({ error: 'activation_failed' }, 500);
-  }
 });
 
 // GET /v1/payments - Get payment history for authenticated user
@@ -218,14 +118,18 @@ paymentRoutes.get('/verify/:trxId', async (c) => {
 
 // GET /v1/payments/stats - Admin stats (requires admin auth)
 paymentRoutes.get('/stats', authMiddleware, adminMiddleware, async (c) => {
+  // Amounts are split by currency: HBD (on-chain) and USD (card) must never be summed
+  // together, or revenue totals mix two units.
   const stats = await db.queryOne(`
     SELECT
       COUNT(*) FILTER (WHERE status = 'processed') as total_payments,
-      SUM(amount) FILTER (WHERE status = 'processed') as total_hbd,
+      SUM(amount) FILTER (WHERE status = 'processed' AND currency = 'HBD') as total_hbd,
+      SUM(amount) FILTER (WHERE status = 'processed' AND currency = 'USD') as total_usd,
       SUM(months_credited) FILTER (WHERE status = 'processed') as total_months,
       COUNT(DISTINCT tenant_id) FILTER (WHERE status = 'processed') as unique_tenants,
       COUNT(*) FILTER (WHERE status = 'processed' AND created_at > NOW() - INTERVAL '30 days') as payments_30d,
-      SUM(amount) FILTER (WHERE status = 'processed' AND created_at > NOW() - INTERVAL '30 days') as hbd_30d
+      SUM(amount) FILTER (WHERE status = 'processed' AND currency = 'HBD' AND created_at > NOW() - INTERVAL '30 days') as hbd_30d,
+      SUM(amount) FILTER (WHERE status = 'processed' AND currency = 'USD' AND created_at > NOW() - INTERVAL '30 days') as usd_30d
     FROM payments
   `);
 
@@ -236,11 +140,13 @@ paymentRoutes.get('/stats', authMiddleware, adminMiddleware, async (c) => {
   return c.json({
     totalPayments: parseInt(stats.total_payments || '0'),
     totalHbd: parseFloat(stats.total_hbd || '0'),
+    totalUsd: parseFloat(stats.total_usd || '0'),
     totalMonthsCredited: parseInt(stats.total_months || '0'),
     uniqueTenants: parseInt(stats.unique_tenants || '0'),
     last30Days: {
       payments: parseInt(stats.payments_30d || '0'),
       hbd: parseFloat(stats.hbd_30d || '0'),
+      usd: parseFloat(stats.usd_30d || '0'),
     },
     activeSubscriptions: parseInt(activeSubscriptions?.count || '0'),
   });
